Sending, receiving, decoding and encoding messages moved to connection.py
