+++ data.json
@@ -1,0 +1,19 @@
+{
+  "Filip": {
+    "username": "Filip",
+    "password": "password",
+    "type": "admin",
+    "inbox": [
+      {
+        "sender": "Agata",
+        "message": "Hello there!"
+      }
+    ]
+  },
+  "Agata": {
+    "username": "Agata",
+    "password": "kotek",
+    "type": "user",
+    "inbox": []
+  }
+}
