+++ databse.py
@@ -1,0 +1,8 @@
+# database.py
+
+class Database:
+
+    def __init__(self):
+
+        pass
+
