+++ query.py
@@ -1,0 +1,9 @@
+# query.py
+
+from databse import Database
+
+class Query:
+
+    def __init__(self):
+
+        pass
